modified both api and webservice

+++ weather-app-webservice/src/main/java/com/app/weather/webservice/payload/WeatherInfo.java
@@ -3,8 +3,6 @@ package com.app.weather.webservice.payload;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigInteger;
-
 @Getter
 @Setter
 public class WeatherInfo {
@@ -12,6 +10,6 @@ public class WeatherInfo {
     private int type;
     private double message;
     private String country;
-    private BigInteger sunrise;
-    private BigInteger sunset;
+    private Long sunrise;
+    private Long sunset;
 }
